Change VP elem compute key persistence
Vertex program execution on Spark (specifically
SparkExecutor.executeVertexProgramIteration) assumes that vertices
inside the input graphRDD can be modified in-place.  In-place
modification of RDD data is generally not allowed; it works with some
combinations of concrete RDD types and rdd.cache(), but not all.  For
instance, if graphRDD is a UnionRDD, in-place modifications to vertex
element compute key properties are lost, even if cache() is called.
If graphRDD is a ShuffledRDD (e.g. as a result of calling reduceByKey
or partitionBy), then it seems like cache() allows subsequent in-place
modification.

This commit changes the code that writes the graphRDD to disk after
the vertex program computes.  It now combines the graphRDD with a
separate RDD (built by SparkExecutor and named "viewIncomingRDD")
containing the final iteration's element compute keys.  It does this
by joining graphRDD and viewIncomingRDD via outer join by vertex IDs,
then running a map() to attaching element compute key properties to
those vertices that had them on the final iteration and emit just the
vertex writables as values.  The resulting vertex writables are then
written in the same way as graphRDD's values were written before this
commit.